added search icon and fixed some issues in searchAPI
