+++ src/App.css
@@ -1,0 +1,7 @@
+a {
+    text-decoration: none;
+}
+
+textarea {
+    resize: none;
+}
