+++ src/geoAPI/CitiesByCountry.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+function CitiesByCountry(props) {
+  return <div></div>;
+}
+
+export default CitiesByCountry;
