+++ src/geoAPI/PopResult.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+function PopResult(props) {
+  return <div></div>;
+}
+
+export default PopResult;
